backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-multipart>=0.0.9
tensorflow>=2.16.1
opencv-python>=4.10.0.84
pillow>=10.3.0
numpy>=1.26.4
sqlalchemy>=2.0.31
alembic>=1.13.2
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
